Add support for reading / writing EEPROM
With this update the length can be larger than 256 (words). We used to
ignore the high byte of the length (and the type for that matter).

Before we were always wring SPM_PAGESIZE/2 words, but now we actually
obey the given length.

This update brings it very close to the "official" version 6.0. There
are only small changes, but the commits are much smaller than the one
big 6.0 update.